requirements.txt
fastapi
uvicorn
anthropic
python-dotenv
pydantic
httpx
jinja2
python-multipart
pillow
python-jose[cryptography]
passlib[bcrypt]
stripe
sqlalchemy
alchemical
alembic
psycopg2-binary
google-auth
email-validator
google-generativeai
celery[redis]
redis
slowapi
gunicorn
